move email helper to utils
